requirements.txt
cryptography>=41.0.0
numpy>=1.24.0
pytest>=7.0.0
PySocks>=1.7.1
requests>=2.28.0
beautifulsoup4>=4.11.0
matplotlib>=3.6.0
psutil>=5.9.0
imagehash>=4.3.0
Pillow>=9.0.0
reportlab>=3.6.0
gpt4all>=2.0.0